asan_symbolize: Improve file path resolution with sysroot enabled
The script supports sysroots, but isn't very smart about how those are handled:
Basically every path that is encountered in the log is prepended with the
sysroot directory. This is OK for libraries, but breaks symbolization in the
binary itself - since that is usually in the build directory. Thus, check first
if file is present in sysroot and - if not - return the original path.

Issue: <INTERNAL>